Add singleBlock method to Parallel class
A new method named 'singleBlock' has been added to the Parallel class. This method allows tasks to be executed once across multiple threads, evidenced with the AtomicBoolean 'executed' instance. Correspondingly, a new test case 'testSingle' has been added to the 'ParallelTests' class to validate the behavior of the new feature.